MAJOR changes. multiple 0 followers, no more "all proj" view
Now each project will have an item that follows 0 and is a child of 0.
This provides unambiguous starting point for hoist and other operations.
Changed routines like getLastChild to be relative to a give project.
Remember, you only need to know the last item that needs to be followed,
you don't need to know the specific benk ID - benk will create that for you.